api/serializers.py
from rest_framework import serializers

from .models import Notebook, Note


class NotebookSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Notebook
        fields = "__all__"


class NoteSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Note
        fields = "__all__"
