requirements.txt
Django==2.1
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
gunicorn==19.9.0
pytz==2018.5
sqlparse==0.2.4
